+++ back-end/src/main/java/com/br/api/dto/aluno/AlunoCadastroDTO.java
@@ -30,5 +30,5 @@ public record AlunoCadastroDTO(
         String email,
 
         @NotNull(message = "ID da turma é obrigatório")
-        int turmaId
+        Long turmaId
 ) {  }
