+++ back-end/src/main/java/com/br/api/repository/TurmaRepository.java
@@ -9,9 +9,9 @@ import com.br.api.model.Turma;
 
 
 @Repository
-public interface TurmaRepository extends JpaRepository<Turma, Integer> {
+public interface TurmaRepository extends JpaRepository<Turma, Long> {
     List<Turma> findByTurno(String turno);
 
     @Query("SELECT t FROM Turma t JOIN t.professores p WHERE p.id_professor = :professorId")
-    List<Turma> findByProfessorId(int professorId);
+    List<Turma> findByProfessorId(Long professorId);
 }
